Unify `buildSchemaFromSDL` and `buildSchemaFromModules`
