+++ packages/apollo-graphql/src/schema/__tests__/buildSchemaFromSDL.test.ts
@@ -1,6 +1,11 @@
 import gql from "graphql-tag";
 import { buildSchemaFromSDL } from "../buildSchemaFromSDL";
-import { GraphQLSchema, GraphQLDirective, DirectiveLocation } from "graphql";
+import {
+  GraphQLSchema,
+  GraphQLDirective,
+  DirectiveLocation,
+  GraphQLObjectType
+} from "graphql";
 
 import astSerializer from "./snapshotSerializers/astSerializer";
 import graphQLTypeSerializer from "./snapshotSerializers/graphQLTypeSerializer";
@@ -338,6 +343,34 @@ type MutationRoot {
       ).toThrowErrorMatchingInlineSnapshot(
         `"Unknown directive \\"something\\"."`
       );
+    });
+  });
+  describe(`resolvers`, () => {
+    it(`should add a resolver for a field`, () => {
+      const name = () => {};
+
+      const schema = buildSchemaFromSDL([
+        {
+          typeDefs: gql`
+            type User {
+              name: String
+            }
+          `,
+          resolvers: {
+            User: {
+              name
+            }
+          }
+        }
+      ]);
+
+      const userType = schema.getType("User");
+      expect(userType).toBeDefined();
+
+      const nameField = (userType! as GraphQLObjectType).getFields()["name"];
+      expect(nameField).toBeDefined();
+
+      expect(nameField.resolve).toEqual(name);
     });
   });
 });
